Feature : formatPrice 함수 분리
