+++ src/basic/App.tsx
@@ -17,6 +17,7 @@ import {
 import { AppLayout } from "./components/layout/AppLayout";
 import { HeaderLayout } from "./components/layout/HeaderLayout";
 import { NotificationToast } from "./components/common/NotificationToast";
+import { formatPrice } from "./utils/format";
 
 const App = () => {
   const [products, setProducts] = useState<ProductWithUI[]>(() => {
@@ -82,21 +83,6 @@ const App = () => {
     discountType: "amount" as "amount" | "percentage",
     discountValue: 0,
   });
-
-  const formatPrice = (price: number, productId?: string): string => {
-    if (productId) {
-      const product = products.find((p) => p.id === productId);
-      if (product && getRemainingStock(product) <= 0) {
-        return "SOLD OUT";
-      }
-    }
-
-    if (isAdmin) {
-      return `${price.toLocaleString()}원`;
-    }
-
-    return `₩${price.toLocaleString()}`;
-  };
 
   const getMaxApplicableDiscount = (item: CartItem): number => {
     const { discounts } = item.product;
@@ -519,48 +505,52 @@ const App = () => {
                       </tr>
                     </thead>
                     <tbody className="bg-white divide-y divide-gray-200">
-                      {(activeTab === "products" ? products : products).map(
-                        (product) => (
-                          <tr key={product.id} className="hover:bg-gray-50">
-                            <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                              {product.name}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                              {formatPrice(product.price, product.id)}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                              <span
-                                className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                                  product.stock > 10
-                                    ? "bg-green-100 text-green-800"
-                                    : product.stock > 0
-                                    ? "bg-yellow-100 text-yellow-800"
-                                    : "bg-red-100 text-red-800"
-                                }`}
-                              >
-                                {product.stock}개
-                              </span>
-                            </td>
-                            <td className="px-6 py-4 text-sm text-gray-500 max-w-xs truncate">
-                              {product.description || "-"}
-                            </td>
-                            <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                              <button
-                                onClick={() => startEditProduct(product)}
-                                className="text-indigo-600 hover:text-indigo-900 mr-3"
-                              >
-                                수정
-                              </button>
-                              <button
-                                onClick={() => deleteProduct(product.id)}
-                                className="text-red-600 hover:text-red-900"
-                              >
-                                삭제
-                              </button>
-                            </td>
-                          </tr>
-                        )
-                      )}
+                      {products.map((product) => (
+                        <tr key={product.id} className="hover:bg-gray-50">
+                          <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
+                            {product.name}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                            {formatPrice(
+                              products,
+                              isAdmin,
+                              product.price,
+                              getRemainingStock,
+                              product.id
+                            )}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                            <span
+                              className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
+                                product.stock > 10
+                                  ? "bg-green-100 text-green-800"
+                                  : product.stock > 0
+                                  ? "bg-yellow-100 text-yellow-800"
+                                  : "bg-red-100 text-red-800"
+                              }`}
+                            >
+                              {product.stock}개
+                            </span>
+                          </td>
+                          <td className="px-6 py-4 text-sm text-gray-500 max-w-xs truncate">
+                            {product.description || "-"}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                            <button
+                              onClick={() => startEditProduct(product)}
+                              className="text-indigo-600 hover:text-indigo-900 mr-3"
+                            >
+                              수정
+                            </button>
+                            <button
+                              onClick={() => deleteProduct(product.id)}
+                              className="text-red-600 hover:text-red-900"
+                            >
+                              삭제
+                            </button>
+                          </td>
+                        </tr>
+                      ))}
                     </tbody>
                   </table>
                 </div>
@@ -1064,7 +1054,13 @@ const App = () => {
                             {/* 가격 정보 */}
                             <div className="mb-3">
                               <p className="text-lg font-bold text-gray-900">
-                                {formatPrice(product.price, product.id)}
+                                {formatPrice(
+                                  products,
+                                  isAdmin,
+                                  product.price,
+                                  getRemainingStock,
+                                  product.id
+                                )}
                               </p>
                               {product.discounts.length > 0 && (
                                 <p className="text-xs text-gray-500">
